fixed open app error
When app opens the numberOfPages Field is set to null. Had to account for this error. 

However the error is still not fixed. The app is not displaying the message when pagesRead is greater then number of pages.  This seems to e happening with the Dates fields as well. There issue does not seem to extend from just the new code.